requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.6.4
Markdown==3.3.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0
